fix: fix the docker image parser to account for private repos
