+++ drivers/docker/driver_test.go
@@ -2852,30 +2852,6 @@ func TestDockerDriver_AdvertiseIPv6Address(t *testing.T) {
 	}
 }
 
-func TestParseDockerImage(t *testing.T) {
-	ci.Parallel(t)
-
-	tests := []struct {
-		Image string
-		Repo  string
-		Tag   string
-	}{
-		{"host:5000/library/hello-world", "host:5000/library/hello-world", "latest"},
-		{"host:5000/library/hello-world:1.0", "host:5000/library/hello-world", "1.0"},
-		{"library/hello-world:1.0", "library/hello-world", "1.0"},
-		{"library/hello-world", "library/hello-world", "latest"},
-		{"library/hello-world:latest", "library/hello-world", "latest"},
-		{"library/hello-world@sha256:f5233545e43561214ca4891fd1157e1c3c563316ed8e237750d59bde73361e77", "library/hello-world@sha256:f5233545e43561214ca4891fd1157e1c3c563316ed8e237750d59bde73361e77", ""},
-	}
-	for _, test := range tests {
-		t.Run(test.Image, func(t *testing.T) {
-			repo, tag := parseDockerImage(test.Image)
-			must.Eq(t, test.Repo, repo)
-			must.Eq(t, test.Tag, tag)
-		})
-	}
-}
-
 func TestDockerImageRef(t *testing.T) {
 	ci.Parallel(t)
 	tests := []struct {
